Rename facade02utilitymethods.py
The naming is a bit confusing, using normal module name make it clearer
