+++ augur/tasks/git/util/facade_worker/facade_worker/facade00mainprogram.py
@@ -30,7 +30,7 @@ import traceback
 import sys, platform, imp, time, datetime, html.parser, subprocess, os, getopt, xlsxwriter, configparser, logging
 from multiprocessing import Process, Queue
 from .config import FacadeSession as FacadeSession
-from .facade02utilitymethods import trim_commit, store_working_author, trim_author   
+from .utilitymethods import trim_commit, store_working_author, trim_author   
 from .facade03analyzecommit import analyze_commit
 from .facade04postanalysiscleanup import git_repo_cleanup
 from .facade05repofetch import git_repo_initialize, check_for_repo_updates, force_repo_updates, force_repo_analysis, git_repo_updates
